Move xcodebuild error handling logic to errors.go file
