+++ step/archive.go
@@ -1,12 +1,9 @@
 package step
 
 import (
-	"bufio"
 	"bytes"
-	"errors"
 	"fmt"
 	"os"
-	"os/exec"
 	"strings"
 	"time"
 
@@ -56,43 +53,4 @@ func runArchiveCommand(archiveCmd *xcodebuild.CommandBuilder, useXcpretty bool, 
 	out := output.String()
 
 	return output.String(), wrapXcodebuildCommandError(archiveCmd, out, err)
-}
-
-type Printable interface {
-	PrintableCmd() string
-}
-
-func wrapXcodebuildCommandError(cmd Printable, out string, err error) error {
-	if err == nil {
-		return nil
-	}
-
-	var exitErr *exec.ExitError
-	if errors.As(err, &exitErr) {
-		reasons := findErrors(out)
-		if len(reasons) > 0 {
-			return fmt.Errorf("command (%s) failed with exit status %d: %w", cmd.PrintableCmd(), exitErr.ExitCode(), errors.New(strings.Join(reasons, "\n")))
-		}
-		return fmt.Errorf("command (%s) failed with exit status %d", cmd.PrintableCmd(), exitErr.ExitCode())
-	}
-
-	return fmt.Errorf("executing command (%s) failed: %w", cmd.PrintableCmd(), err)
-}
-
-func findErrors(out string) []string {
-	var errors []string
-
-	scanner := bufio.NewScanner(strings.NewReader(out))
-	scanner.Split(bufio.ScanLines)
-	for scanner.Scan() {
-		line := scanner.Text()
-		if strings.HasPrefix(line, "error: ") {
-			errors = append(errors, line)
-		}
-	}
-	if err := scanner.Err(); err != nil {
-		return nil
-	}
-
-	return errors
 }
